I've refactored the `MistweaverMonk.lua` script to improve its performance, readability, and maintainability.
The main changes are:

- I centralized all unit scanning logic into a single `UnitScanner` class to iterate over units only once per frame, which significantly improves performance.
- I refactored the APLs to use the new `UnitScanner` class, making them more concise and easier to understand.
- I replaced global slash commands with `Bastion.Command` to avoid polluting the global namespace.
- I removed redundant helper functions and custom units.
- I merged all helper code into a single file to resolve file loading issues in the game client.
